refactor: extract orchestration trait interfaces to orchestration-domain entity (Phase 3a)
Move AgentPoolManager/PoolError/SlotHandle, WorkflowService/WorkflowServiceError,
AgentRegistry, and TaskExecutor trait definitions to orchestration-domain entity
package. Live implementations remain in root orchestration.control with renamed
companion objects (AgentPoolManagerLive, AgentRegistryLive, WorkflowServiceLive,
TaskExecutorLive). Update 39 consumer files to import from orchestration.entity.
