+++ src/main/scala/config/boundary/AgentsController.scala
@@ -14,7 +14,7 @@ import agent.control.{ AgentMatching, BuiltInAgentSynchronizer }
 import agent.entity.api.*
 import agent.entity.{ Agent as RegistryAgent, AgentEvent, AgentPermissions, AgentRepository, TrustLevel }
 import llm4zio.core.{ LlmError, LlmService }
-import orchestration.control.AgentRegistry
+import orchestration.entity.AgentRegistry
 import orchestration.entity.AgentMonitorService
 import prompts.PromptLoader
 import shared.errors.PersistenceError
