+++ src/main/scala/gateway/control/WorkflowNotifier.scala
@@ -6,7 +6,8 @@ import zio.json.*
 import _root_.config.entity.WorkflowRow
 import db.*
 import gateway.entity.*
-import orchestration.control.{ AgentRegistry, TaskExecutor }
+import orchestration.entity.TaskExecutor
+import orchestration.entity.AgentRegistry
 import shared.errors.PersistenceError
 import taskrun.entity.{ RunStatus, TaskRunRow }
