+++ src/main/scala/orchestration/control/Llm4zioAdapters.scala
@@ -5,6 +5,7 @@ import java.time.Instant
 import zio.*
 
 import _root_.config.entity.AgentInfo
+import orchestration.entity.AgentRegistry
 import conversation.entity.api.{ ChatConversation, ConversationEntry, MessageType, SenderType }
 import db.ChatRepository
 import gateway.entity.SessionKey
