+++ src/main/scala/project/boundary/ProjectsController.scala
@@ -13,7 +13,7 @@ import analysis.control.WorkspaceAnalysisScheduler
 import analysis.entity.{ WorkspaceAnalysisState, WorkspaceAnalysisStatus }
 import issues.entity.api.{ AgentIssueView, IssuePriority, IssueStatus }
 import issues.entity.{ AgentIssue, IssueFilter, IssueRepository, IssueState }
-import orchestration.control.AgentRegistry
+import orchestration.entity.AgentRegistry
 import project.control.ProjectStorageService
 import project.entity.*
 import shared.errors.PersistenceError
