+++ src/test/scala/app/control/HealthMonitorSpec.scala
@@ -10,7 +10,7 @@ import _root_.config.entity.{ CustomAgentRow, SettingRow, * }
 import db.*
 import gateway.control.*
 import gateway.entity.{ GatewayMetricsSnapshot, NormalizedMessage, SessionKey, SessionScopeStrategy }
-import orchestration.control.AgentRegistry
+import orchestration.entity.AgentRegistry
 import shared.errors.PersistenceError
 import taskrun.entity.{ RunStatus, TaskArtifactRow, TaskReportRow, TaskRunRow, TaskStep }
