+++ src/test/scala/workspace/boundary/WorkspacesControllerSpec.scala
@@ -11,7 +11,7 @@ import _root_.config.entity.*
 import analysis.control.WorkspaceAnalysisScheduler
 import analysis.entity.{ AnalysisType, WorkspaceAnalysisState, WorkspaceAnalysisStatus }
 import issues.entity.{ AgentIssue, IssueEvent, IssueFilter, IssueRepository, IssueState }
-import orchestration.control.AgentRegistry
+import orchestration.entity.AgentRegistry
 import shared.ids.Ids.IssueId
 import taskrun.entity.TaskStep
 import workspace.control.{ GitService, WorkspaceRunService }
